# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TRIMESTRA API",
        "description": "API Documentation for TRIMESTRA - Pregnancy Nutrition App",
        "contact": {
            "email": "support@trimestra.test"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://trimestra-backend.test/api/v1",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/foods/{id}": {
            "get": {
                "tags": [
                    "Foods"
                ],
                "summary": "Get food by ID",
                "operationId": "000eab66c08730d6dcb32b9386cd8d5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Food ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Food retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Food"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Food retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Food not found"
                    }
                }
            }
        },
        "/foods/category/{categoryId}": {
            "get": {
                "tags": [
                    "Foods"
                ],
                "summary": "Get foods by category",
                "operationId": "e36953ae08cf8019b0e9a4a2a5aa671d",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foods by category retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Food"
                                                    }
                                                },
                                                "links": {
                                                    "type": "object"
                                                },
                                                "meta": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Foods by category retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/foods/recommended": {
            "get": {
                "tags": [
                    "Foods"
                ],
                "summary": "Get recommended foods for user",
                "operationId": "0e0754f895f7a05a62cd95100e83e016",
                "responses": {
                    "200": {
                        "description": "Recommended foods retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "foods": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Food"
                                                    }
                                                },
                                                "total_results": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Recommended foods retrieved successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Food": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "calories": {
                        "type": "number",
                        "format": "float"
                    },
                    "protein": {
                        "type": "number",
                        "format": "float"
                    },
                    "carbs": {
                        "type": "number",
                        "format": "float"
                    },
                    "fat": {
                        "type": "number",
                        "format": "float"
                    },
                    "fiber": {
                        "type": "number",
                        "format": "float"
                    },
                    "serving_size": {
                        "type": "string"
                    },
                    "cooking_time": {
                        "type": "integer"
                    },
                    "is_pregnancy_safe": {
                        "type": "boolean"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "allergens": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "diet_types": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}